requirements.txt
# local package
-e .

# external requirements
# click
# Sphinx
# coverage
# awscli
# flake8
# python-dotenv>=0.5.1

pandas
numpy
scikit-learn
PyYAML
nltk
lightgbm
mlflow
matplotlib
seaborn
boto3
